+++ lib/src/file.dart
@@ -1,9 +1,10 @@
+library box.file;
+
 import 'dart:convert';
 import 'dart:io';
 
 import 'package:box/core.dart';
 import 'package:box/src/memory.dart';
-import 'package:reflective/reflective.dart';
 
 class FileBox extends MemoryBox {
   final String _path;
@@ -12,28 +13,27 @@ class FileBox extends MemoryBox {
   @override
   bool get persistent => true;
 
-  FileBox(this._path);
+  FileBox(this._path, Registry registry) : super(registry);
 
+  @override
   Future store(Object entity) {
     super.store(entity);
-    return Future(() => _persist(TypeReflection.fromInstance(entity)));
+    return Future(() => _persist(entity.runtimeType));
   }
 
-  Future _persist(TypeReflection type) async {
+  Future _persist(Type type) async {
     if (_persisting) {
       // TODO: stage changes
       throw 'Already persisting changes. Please use await box.store(...) to make sure only one change is persisted at a time';
     }
     _persisting = true;
     var entities = await entitiesFor(type);
-    var file = _fileOf(type.name);
+    var file = _fileOf(registry.lookup(type).name);
     if (file.existsSync()) {
       await file.delete();
     }
     await file.create(recursive: true);
-    var json = await Stream.fromIterable(entities.values)
-        .map((value) => jsonEncode(Conversion.convert(value).to(Map)).toString())
-        .join("\n");
+    var json = await Stream.fromIterable(entities.values).map((value) => jsonEncode(value)).join('\n');
     await file.writeAsString(json.toString(), mode: FileMode.append);
     _persisting = false;
   }
@@ -43,24 +43,26 @@ class FileBox extends MemoryBox {
   }
 
   @override
-  Future<Map> entitiesFor(TypeReflection type) {
-    entities.putIfAbsent(type.name, () => Map());
-    if (entities[type.name].isEmpty) {
+  Future<Map> entitiesFor(Type type) {
+    var typeName = registry.lookup(type).name;
+    entities.putIfAbsent(typeName, () => {});
+    if (entities[typeName].isEmpty) {
       return _load(type).toList().then((values) {
-        entities[type.name] = Maps.index(values, (value) => Box.keyOf(value));
-        return entities[type.name];
+        entities[typeName] = {for (var value in values) keyOf(value): value};
+        return entities[typeName];
       });
     }
-    return Future.value(entities[type.name]);
+    return Future.value(entities[typeName]);
   }
 
-  Stream _load(TypeReflection reflection) {
-    File file = _fileOf(reflection.name);
+  Stream _load(Type type) {
+    var entitySupport = registry.lookup(type);
+    var file = _fileOf(entitySupport.name);
     return Stream.fromFuture(file.exists().then((exists) {
       if (exists) {
         return file.openRead().transform(utf8.decoder).transform(const LineSplitter()).map((line) {
-          if (line.startsWith("{")) {
-            return Conversion.convert(jsonDecode(line)).to(reflection.rawType);
+          if (line.startsWith('{')) {
+            return entitySupport.deserialize(jsonDecode(line));
           }
           return null;
         }).where((item) => item != null);
@@ -69,23 +71,22 @@ class FileBox extends MemoryBox {
     })).asyncExpand((Stream stream) => stream);
   }
 
+  @override
   Future<T> find<T>(key, [Type type]) {
-    TypeReflection reflection = TypeReflection<T>(type);
-    String typeName = reflection.name;
+    var typeName = registry.lookup(type ?? T).name;
     if (entities.isEmpty) {
-      return _load(reflection).toList().then((values) {
-        entities[typeName] = Maps.index(values, (value) => Box.keyOf(value));
-        return super.find<T>(key, type);
+      return _load(type ?? T).toList().then((values) {
+        entities[typeName] = {for (var value in values) keyOf(value): value};
+        return super.find<T>(key, type ?? T);
       });
     }
-    return super.find<T>(key, type);
+    return super.find<T>(key, type ?? T);
   }
 
   @override
   Future deleteAll<T>([Type type]) async {
     await super.deleteAll<T>(type);
-    var reflection = TypeReflection<T>(type);
-    var file = _fileOf(reflection.name);
+    var file = _fileOf(registry.lookup(type ?? T).name);
     if (file.existsSync()) {
       await file.delete();
     }